Major changes to the download function
I used ChatGPT to comment and revise the code of this function.
I added a lot of new things, including the integration between the select_quarter function and the download_quarter one.
Also, I found a function that puts the "0" before the quarter number, facilitating on the working process of the select_quarter function.