Deleting the old function's code and making changes on the new code
I had kept the old function's code, for now, as i like to build my new cod always checking for the old one, this commit should be considered an ordinary one, more as a work routine specifically for me.
I also made significant change son the new code, adding a bigger timeout for downloading the file, in order to handle slower connections, I added a old function at the end that I had somewhat forgotten to add when I deleted the old code. I added code chunks that put the file in a subfolder inside "PNADC_Microdata" with the name of the year of that PNAD and another one that takes the downloaded file and generates an unzipped version at the same folder.