Repository reallocation - Stata files- -
There was a "Stata" folder in this repository, but, it looked fairly old, with the new PNAD codes for Stata running perfectly but in another repo of Datazoom. Therefore, I primarily moved this folder and all its files to the new repo, and named it as "Stata_old_repo". Pending consideration, this folder could be deleted as well, please check the "PNAD" Repo's history in order to check if this folder was deleted or modified, what is very likely.