Creating a loop for automating panel-detection on the downloaded files
We created a loop that identifies from what panel each file (that was downloaded in the panel above this one) is from, from it's termination, which indicates exactly each panel thos observations are from.
Eg. the file "2022_2_9" has the observations that are from the year 2022, 2nd quarter, 9th panel.
This loop is able to identify the files from your PC (just remember to check if the working directory is defined correctly, it must be the same which contains the downloaded PNAD files).
I believe this could be integrated with the download loop in the future, so the   user can get either the quarter he desires, or the panel most specifically.