Creating the function bundle_panel
As we can see, this function, if given the right directory, will scan the directory for 8 archives ended in "_x", give that x ranges from 1 to 9 (the actualy numbers of panels (variable V1016 of the PNAD) available in the PNAD.
This function is better used when combined with our download function, which already loads into a directory the archives of all panels wanted.
Therefore, we use the archive-naming model (year_quarter_panel), to read all the RDS and bundle them into data frames.
Then, this function creates a list with every data frame of every panel in the loop (we'll modify it later so the user can select which panels it wants to bundle, or we'll add this function to the download one, as they complement each other).