Creating the function cria_df_de_atrito
This function receives a certain data-frame, it first filters the dataframe (a PNAD panel database) to have only the people who have responded the first interview (which is the one we'll use as a comparison base).
After filtering, it stores, for each person, two vectors, inside 2 cells of another dataframe (this was done using the summarize function), the first vectors, called appearances, records which interviews the individual was present (as we can id the individual using our identificator (id_ind), developed before in this package), and the disappearances does the exact opposite.
After that, the function creates 5 dummy columns, which acquire the value of 1 if the person missed the 1st interview (for the 1st new column), an so on for every interview.
After that, we sum each column to record the number of missing persons on every interview and do some calculations to return the "atrito_definite" dataframe as a result, which has 3 columns: 1st one: Just a index of what interview that line refers to, 2nd one: The number of missing persons in every interview, 3rd one: The percentage of people FOUND in every interview.