Updating and renaming the download_years function
I changed a series of things in the download_years function. Firstly, I changed its name (as I have also done in the code that is on the "functions_improvements" branch) to load_pnad. In second place, I also did add the "quarters" parameter, to indicate to users that they can now specify the quarters they'd like to download as well as the years.