Re-arranging the READ.ME
I rearranged the READ.ME, as instructed by @pablotadeu , so the load_pnadc function could get more focus and attention from the reader, as it is the main function of this package.
The new order is: load_pnadc -> bundle_panel -> load_pnadc_panel -> all other support functions of the package.