Adding the friction calculation file
This file was organized and adapted from an older one, that was called "Calculo_atrito_efetivo_29_08_23.R" and was deleted in commit 2c7ef9bd3f6afa184d443fb6bec9c2018f96b628 .
I adapted and documented it to this function.
It is not fully complete, as its application could be better adapted to be applied to multiple panel files.
This should be an internal function at first.