Adjusting the code to fit the advanced panel
Here I've inserted the structure necessary to run the advanced panel identification algorithm, which applies only to the yet-to-be fully-matched observations.
To identify the "to-be-treated" observations, we must first run the basic panel identification algorithm, and then identify them.
After this identification process, we can implement the advanced panel algorithm, which will be later specified in PR #40 , where we develop the function "advanced_panel_1st_level" shown in this commit's code.