Adjusting the advanced panel algorithm
In the application of the advanced panel algorithm's, we noticed that we could be applying the cur_group_id function all over again to the non-fully-matched observations.
Therefore, I adjusted this by setting the start level for the new id's created by cur_group_id, which is m, the "biggest" id created in the basic panel algorithm.